+++ hoot-core/src/main/cpp/hoot/core/algorithms/rubber-sheet/RubberSheetApplier.cpp
@@ -33,7 +33,7 @@
 #include <hoot/core/util/ConfigOptions.h>
 #include <hoot/core/util/HootException.h>
 #include <hoot/core/util/Log.h>
-#include <hoot/core/util/IoUtils.h>
+#include <hoot/core/io/IoUtils.h>
 #include <hoot/core/algorithms/rubber-sheet/RubberSheet.h>
 #include <hoot/core/ops/MapCleaner.h>
 #include <hoot/core/util/Settings.h>
