+++ hoot-core/src/main/cpp/hoot/core/cmd/DiffCmd.cpp
@@ -33,7 +33,7 @@
 #include <hoot/core/util/Settings.h>
 #include <hoot/core/elements/OsmMap.h>
 #include <hoot/core/util/Log.h>
-#include <hoot/core/util/IoUtils.h>
+#include <hoot/core/io/IoUtils.h>
 
 using namespace std;
