+++ hoot-core/src/main/cpp/hoot/core/cmd/TagCompareCmd.cpp
@@ -29,7 +29,7 @@
 #include <hoot/core/util/Factory.h>
 #include <hoot/core/cmd/BaseCommand.h>
 #include <hoot/core/schema/AttributeCoOccurrence.h>
-#include <hoot/core/util/IoUtils.h>
+#include <hoot/core/io/IoUtils.h>
 
 using namespace std;
