+++ hoot-core/src/main/cpp/hoot/core/io/ApiDbReader.cpp
@@ -33,7 +33,7 @@
 #include <hoot/core/io/ApiDb.h>
 #include <hoot/core/util/Log.h>
 #include <hoot/core/util/StringUtils.h>
-#include <hoot/core/util/IoUtils.h>
+#include <hoot/core/io/IoUtils.h>
 
 // tgs
 #include <tgs/System/Time.h>
