+++ hoot-core/src/main/cpp/hoot/core/io/IoUtils.h
