+++ hoot-core/src/main/cpp/hoot/core/io/MapStatsWriter.cpp
@@ -31,7 +31,7 @@
 #include <hoot/core/info/SingleStat.h>
 #include <hoot/core/elements/OsmMap.h>
 #include <hoot/core/util/Factory.h>
-#include <hoot/core/util/IoUtils.h>
+#include <hoot/core/io/IoUtils.h>
 #include <hoot/core/util/MapProjector.h>
 #include <hoot/core/util/Settings.h>
 #include <hoot/core/util/ConfigOptions.h>
