+++ hoot-core/src/main/cpp/hoot/core/io/OsmJsonReader.cpp
@@ -35,7 +35,7 @@
 #include <hoot/core/util/HootException.h>
 #include <hoot/core/util/Log.h>
 #include <hoot/core/util/StringUtils.h>
-#include <hoot/core/util/IoUtils.h>
+#include <hoot/core/io/IoUtils.h>
 
 // Boost
 #include <boost/foreach.hpp>
