+++ hoot-core/src/main/cpp/hoot/core/scoring/RandomForestModelBuilder.cpp
@@ -35,7 +35,7 @@
 #include <hoot/core/scoring/MatchFeatureExtractor.h>
 #include <hoot/core/util/ConfigOptions.h>
 #include <hoot/core/util/Log.h>
-#include <hoot/core/util/IoUtils.h>
+#include <hoot/core/io/IoUtils.h>
 #include <hoot/core/util/StringUtils.h>
 
 #include <hoot/core/io/ArffToRfConverter.h>
