+++ hoot-rnd/src/main/cpp/hoot/rnd/cmd/MultiaryPoiConflateCmd.cpp
@@ -40,7 +40,7 @@
 #include <hoot/core/util/ConfigOptions.h>
 #include <hoot/core/util/Log.h>
 #include <hoot/rnd/conflate/multiary/MultiaryUtilities.h>
-#include <hoot/core/util/IoUtils.h>
+#include <hoot/core/io/IoUtils.h>
 
 // Qt
 #include <QFileInfo>
