+++ hoot-rnd/src/main/cpp/hoot/rnd/cmd/MultiaryScorePoiMatchesCmd.cpp
@@ -39,7 +39,7 @@
 #include <hoot/core/util/MapProjector.h>
 #include <hoot/core/visitors/CalculateHashVisitor.h>
 #include <hoot/rnd/conflate/multiary/MultiaryUtilities.h>
-#include <hoot/core/util/IoUtils.h>
+#include <hoot/core/io/IoUtils.h>
 
 using namespace std;
 using namespace Tgs;
